feat: implement multi vehicle for ant colonies
