+++ Livrable 2/Algorithms/Greedy/GreedyAlgo.py
