--- Livrable 2/AntColonies/Enums/GraphType.py
@@ -1,2 +1,0 @@
-from enum import Enum
-GraphType = Enum('GraphType', [('Random', 1), ('Map', 2)])
